+++ src/pages/LeadDetails.js
@@ -3,32 +3,23 @@ import { useParams, useNavigate } from 'react-router-dom';
 import NavBar from '../components/NavBar';
 import { Grid, Paper } from '@mui/material';
 import { makeStyles, Container } from '@material-ui/core';
-import LeadFormRoachAssessment from '../components/forms/LeadFormRoachAssessment';
+import LeadForm from '../components/forms/LeadForm';
 import axios from 'axios';
 import moment from 'moment';
 
 
 const useStyles = makeStyles(theme => ({
     pageContent: {
-        // margin: theme.spacing(3),
         padding: theme.spacing(3)
     }
 }));
-
-function LeadForm ({ formData }) {
-    if (formData.service.toLowerCase() === 'roach assessment') {
-        return <LeadFormRoachAssessment formData={formData} />;
-    } else {
-        return <div>Unknown Service Type</div>;
-    }
-};
 
 function LeadDetails () {
 
     // we'll use id in url param to form requests to API
     const { id } = useParams();
     const [formData, setFormData] = useState(null);
-    
+
     useEffect(async () => {
         await axios.get(`https://smartpestapi.wn.r.appspot.com/lead/${id}`)
         .then(response => {
@@ -54,7 +45,6 @@ function LeadDetails () {
                         <Grid container>
                             <Grid item xs={8}>
                                 <h2>Lead Details</h2>
-                                {/* <h3>Status: {formData.status}</h3> */}
                             </Grid>
                             <Grid item xs={4} align="right" sx={{ pt: 1, pb: 1 }}>
                                 <div><strong>Date Created:</strong> {formData.dateCreated}</div>
@@ -63,7 +53,7 @@ function LeadDetails () {
                             </Grid>
                         </Grid>
                         <Paper className={classes.pageContent}>
-                            {formData ? <LeadForm formData={formData} /> : <div>...loading</div>}
+                            <LeadForm formData={formData} />
                         </Paper>
                     </Container>
                 ) : <div>loading...</div>
